moved data files to datamanagers
