+++ src/components/datamanagers/Colors.js
