+++ src/components/datamanagers/Randomizer.js
